Refactored various ANSI functions into common places.
* Trimmed ansi/functions to remove some overlap with terminal
* String op<< now writes to a default terminal - terminal is now
  in complete control of output.
* Factored out some of terminal's helpers to the detail namespace in
  preparation for more helper functions that will come along.
