+++ test/ansi_functions_test.cpp
@@ -10,89 +10,49 @@ public :
         CPPUNIT_TEST(set_7bit_controls_outputs_correct_sequence);
         CPPUNIT_TEST(set_8bit_controls_outputs_correct_sequence);
         CPPUNIT_TEST(full_reset_outputs_correct_sequence);
-        
+
         CPPUNIT_TEST(double_height_line_top_half_outputs_correct_sequence);
         CPPUNIT_TEST(double_height_line_bottom_half_outputs_correct_sequence);
         CPPUNIT_TEST(single_width_line_outputs_correct_sequence);
         CPPUNIT_TEST(double_width_line_outputs_correct_sequence);
-        
+
         CPPUNIT_TEST(select_default_charset_outputs_correct_sequence);
         CPPUNIT_TEST(select_utf8_charset_outputs_correct_sequence);
         CPPUNIT_TEST(designate_g0_charset_outputs_correct_sequences);
         CPPUNIT_TEST(designate_g1_charset_outputs_correct_sequences);
         CPPUNIT_TEST(designate_g2_charset_outputs_correct_sequences);
         CPPUNIT_TEST(designate_g3_charset_outputs_correct_sequences);
-        
-        CPPUNIT_TEST(move_cursor_up_1_skips_amount);
-        CPPUNIT_TEST(move_cursor_up_2_outputs_correct_sequence);
-        CPPUNIT_TEST(move_cursor_down_1_skips_amount);
-        CPPUNIT_TEST(move_cursor_down_2_outputs_correct_sequence);
-        CPPUNIT_TEST(move_cursor_forward_1_skips_amount);
-        CPPUNIT_TEST(move_cursor_forward_2_outputs_correct_sequence);
-        CPPUNIT_TEST(move_cursor_backward_1_skips_amount);
-        CPPUNIT_TEST(move_cursor_backward_2_outputs_correct_sequence);
-        CPPUNIT_TEST(move_cursor_to_column_origin_skips_column);
-        CPPUNIT_TEST(move_cursor_to_column_outputs_correct_sequence);
-        CPPUNIT_TEST(move_cursor_position_to_origin_skips_column_and_row);
-        CPPUNIT_TEST(move_cursor_position_to_origin_row_skips_row);
-        CPPUNIT_TEST(move_cursor_position_to_origin_column_skips_column);
-        CPPUNIT_TEST(move_cursor_position_outputs_correct_sequence);
-        CPPUNIT_TEST(hide_cursor_outputs_correct_sequence);
-        CPPUNIT_TEST(show_cursor_outputs_correct_sequence);
-        
-        CPPUNIT_TEST(save_cursor_outputs_correct_sequence);
-        CPPUNIT_TEST(restore_cursor_outputs_correct_sequence);
-        
+
         CPPUNIT_TEST(erase_in_display_below_outputs_correct_sequence);
         CPPUNIT_TEST(erase_in_display_above_outputs_correct_sequence);
         CPPUNIT_TEST(erase_in_display_all_outputs_correct_sequence);
-        
+
         CPPUNIT_TEST(erase_in_line_right_outputs_correct_sequence);
         CPPUNIT_TEST(erase_in_line_left_outputs_correct_sequence);
         CPPUNIT_TEST(erase_in_line_all_outputs_correct_sequence);
     CPPUNIT_TEST_SUITE_END();
-    
+
 private :
     void set_7bit_controls_outputs_correct_sequence();
     void set_8bit_controls_outputs_correct_sequence();
     void full_reset_outputs_correct_sequence();
-    
+
     void double_height_line_top_half_outputs_correct_sequence();
     void double_height_line_bottom_half_outputs_correct_sequence();
     void single_width_line_outputs_correct_sequence();
     void double_width_line_outputs_correct_sequence();
-    
+
     void select_default_charset_outputs_correct_sequence();
     void select_utf8_charset_outputs_correct_sequence();
     void designate_g0_charset_outputs_correct_sequences();
     void designate_g1_charset_outputs_correct_sequences();
     void designate_g2_charset_outputs_correct_sequences();
     void designate_g3_charset_outputs_correct_sequences();
-    
-    void move_cursor_up_1_skips_amount();
-    void move_cursor_up_2_outputs_correct_sequence();
-    void move_cursor_down_1_skips_amount();
-    void move_cursor_down_2_outputs_correct_sequence();
-    void move_cursor_forward_1_skips_amount();
-    void move_cursor_forward_2_outputs_correct_sequence();
-    void move_cursor_backward_1_skips_amount();
-    void move_cursor_backward_2_outputs_correct_sequence();
-    void move_cursor_to_column_origin_skips_column();
-    void move_cursor_to_column_outputs_correct_sequence();
-    void move_cursor_position_to_origin_skips_column_and_row();
-    void move_cursor_position_to_origin_row_skips_row();
-    void move_cursor_position_to_origin_column_skips_column();
-    void move_cursor_position_outputs_correct_sequence();
-    void hide_cursor_outputs_correct_sequence();
-    void show_cursor_outputs_correct_sequence();
-    
-    void save_cursor_outputs_correct_sequence();
-    void restore_cursor_outputs_correct_sequence();
-    
+
     void erase_in_display_below_outputs_correct_sequence();
     void erase_in_display_above_outputs_correct_sequence();
     void erase_in_display_all_outputs_correct_sequence();
-    
+
     void erase_in_line_right_outputs_correct_sequence();
     void erase_in_line_left_outputs_correct_sequence();
     void erase_in_line_all_outputs_correct_sequence();
@@ -104,7 +64,7 @@ void ansi_functions_test_fixture::set_7bit_controls_outputs_correct_sequence()
 {
     std::string expected = "\x1B F";
     std::string result = terminalpp::ansi::set_7bit_controls();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -120,7 +80,7 @@ void ansi_functions_test_fixture::full_reset_outputs_correct_sequence()
 {
     std::string expected = "\x1B""c";
     std::string result = terminalpp::ansi::full_reset();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -128,7 +88,7 @@ void ansi_functions_test_fixture::double_height_line_top_half_outputs_correct_se
 {
     std::string expected = "\x1B#3";
     std::string result = terminalpp::ansi::double_height_line_top_half();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -136,7 +96,7 @@ void ansi_functions_test_fixture::double_height_line_bottom_half_outputs_correct
 {
     std::string expected = "\x1B#4";
     std::string result = terminalpp::ansi::double_height_line_bottom_half();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -144,7 +104,7 @@ void ansi_functions_test_fixture::single_width_line_outputs_correct_sequence()
 {
     std::string expected = "\x1B#5";
     std::string result = terminalpp::ansi::single_width_line();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -152,7 +112,7 @@ void ansi_functions_test_fixture::double_width_line_outputs_correct_sequence()
 {
     std::string expected = "\x1B#6";
     std::string result = terminalpp::ansi::double_width_line();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -160,7 +120,7 @@ void ansi_functions_test_fixture::select_default_charset_outputs_correct_sequenc
 {
     std::string expected = "\x1B%@";
     std::string result = terminalpp::ansi::select_default_charset();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -168,7 +128,7 @@ void ansi_functions_test_fixture::select_utf8_charset_outputs_correct_sequence()
 {
     std::string expected = "\x1B%G";
     std::string result = terminalpp::ansi::select_utf8_charset();
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -178,7 +138,7 @@ void ansi_functions_test_fixture::designate_g0_charset_outputs_correct_sequences
     std::string expected_uk       = "\x1B(A";
     std::string expected_us_ascii = "\x1B(B";
     std::string expected_dec_sgr  = "\x1B(%5";
-    
+
     CPPUNIT_ASSERT_EQUAL(
         expected_dec,
         terminalpp::ansi::designate_g0_charset(
@@ -206,7 +166,7 @@ void ansi_functions_test_fixture::designate_g1_charset_outputs_correct_sequences
     std::string expected_uk       = "\x1B)A";
     std::string expected_us_ascii = "\x1B)B";
     std::string expected_dec_sgr  = "\x1B)%5";
-    
+
     CPPUNIT_ASSERT_EQUAL(
         expected_dec,
         terminalpp::ansi::designate_g1_charset(
@@ -234,7 +194,7 @@ void ansi_functions_test_fixture::designate_g2_charset_outputs_correct_sequences
     std::string expected_uk       = "\x1B*A";
     std::string expected_us_ascii = "\x1B*B";
     std::string expected_dec_sgr  = "\x1B*%5";
-    
+
     CPPUNIT_ASSERT_EQUAL(
         expected_dec,
         terminalpp::ansi::designate_g2_charset(
@@ -262,7 +222,7 @@ void ansi_functions_test_fixture::designate_g3_charset_outputs_correct_sequences
     std::string expected_uk       = "\x1B+A";
     std::string expected_us_ascii = "\x1B+B";
     std::string expected_dec_sgr  = "\x1B+%5";
-    
+
     CPPUNIT_ASSERT_EQUAL(
         expected_dec,
         terminalpp::ansi::designate_g3_charset(
@@ -284,156 +244,12 @@ void ansi_functions_test_fixture::designate_g3_charset_outputs_correct_sequences
             terminalpp::ansi::charset::dec_supplementary_graphics));
 }
 
-void ansi_functions_test_fixture::move_cursor_up_1_skips_amount()
-{
-    std::string expected = "\x1B[A";
-    std::string result = terminalpp::ansi::move_cursor_up();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_up_2_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[2A";
-    std::string result = terminalpp::ansi::move_cursor_up(2);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_down_1_skips_amount()
-{
-    std::string expected = "\x1B[B";
-    std::string result = terminalpp::ansi::move_cursor_down();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_down_2_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[2B";
-    std::string result = terminalpp::ansi::move_cursor_down(2);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_forward_1_skips_amount()
-{
-    std::string expected = "\x1B[C";
-    std::string result = terminalpp::ansi::move_cursor_forward();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_forward_2_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[2C";
-    std::string result = terminalpp::ansi::move_cursor_forward(2);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_backward_1_skips_amount()
-{
-    std::string expected = "\x1B[D";
-    std::string result = terminalpp::ansi::move_cursor_backward();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_backward_2_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[2D";
-    std::string result = terminalpp::ansi::move_cursor_backward(2);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_to_column_origin_skips_column()
-{
-    std::string expected = "\x1B[G";
-    std::string result = terminalpp::ansi::move_cursor_to_column();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_to_column_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[2G";
-    std::string result = terminalpp::ansi::move_cursor_to_column(2);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_position_to_origin_skips_column_and_row()
-{
-    std::string expected = "\x1B[H";
-    std::string result = terminalpp::ansi::move_cursor_to_position(1, 1);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_position_to_origin_row_skips_row()
-{
-    std::string expected = "\x1B[7H";
-    std::string result = terminalpp::ansi::move_cursor_to_position(7, 1);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_position_to_origin_column_skips_column()
-{
-    std::string expected = "\x1B[;7H";
-    std::string result = terminalpp::ansi::move_cursor_to_position(1, 7);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::move_cursor_position_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[7;6H";
-    std::string result = terminalpp::ansi::move_cursor_to_position(7, 6);
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::hide_cursor_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[?25l";
-    std::string result = terminalpp::ansi::hide_cursor();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::show_cursor_outputs_correct_sequence()
-{
-    std::string expected = "\x1B[?25h";
-    std::string result = terminalpp::ansi::show_cursor();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::save_cursor_outputs_correct_sequence()
-{
-    std::string expected = "\x1B""7";
-    std::string result = terminalpp::ansi::save_cursor();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
-void ansi_functions_test_fixture::restore_cursor_outputs_correct_sequence()
-{
-    std::string expected = "\x1B""8";
-    std::string result = terminalpp::ansi::restore_cursor();
-    
-    CPPUNIT_ASSERT_EQUAL(expected, result);
-}
-
 void ansi_functions_test_fixture::erase_in_display_below_outputs_correct_sequence()
 {
     std::string expected = "\x1B[J";
     std::string result = terminalpp::ansi::erase_in_display(
         terminalpp::ansi::erase_display::below);
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -442,7 +258,7 @@ void ansi_functions_test_fixture::erase_in_display_above_outputs_correct_sequenc
     std::string expected = "\x1B[1J";
     std::string result = terminalpp::ansi::erase_in_display(
         terminalpp::ansi::erase_display::above);
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -451,7 +267,7 @@ void ansi_functions_test_fixture::erase_in_display_all_outputs_correct_sequence(
     std::string expected = "\x1B[2J";
     std::string result = terminalpp::ansi::erase_in_display(
         terminalpp::ansi::erase_display::all);
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -460,7 +276,7 @@ void ansi_functions_test_fixture::erase_in_line_right_outputs_correct_sequence()
     std::string expected = "\x1B[K";
     std::string result = terminalpp::ansi::erase_in_line(
         terminalpp::ansi::erase_line::right);
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -469,7 +285,7 @@ void ansi_functions_test_fixture::erase_in_line_left_outputs_correct_sequence()
     std::string expected = "\x1B[1K";
     std::string result = terminalpp::ansi::erase_in_line(
         terminalpp::ansi::erase_line::left);
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
 
@@ -478,6 +294,6 @@ void ansi_functions_test_fixture::erase_in_line_all_outputs_correct_sequence()
     std::string expected = "\x1B[2K";
     std::string result = terminalpp::ansi::erase_in_line(
         terminalpp::ansi::erase_line::all);
-    
+
     CPPUNIT_ASSERT_EQUAL(expected, result);
 }
